Fusion des deux versions de Jeu et modification des imports
