+++ src/entites/Joueur.java
@@ -9,7 +9,7 @@ import ressources.TypeRessource;
 import carte.Direction;
 import batiments.*;
 import exceptions.RessourceInsuffisanteException;
-import ressources.Item;
+import carte.Item;
 
 /**
  * Représente l'entité contrôlée par l'utilisateur
